backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
pydantic
SQLAlchemy
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose
python-multipart
